bin/show-alert.rb: It turns out that we can ignore the SystemExit exceptions here without an unwarranted amount of code.  I have some sympathy with Elliott's argument that we could use exit!() everywhere but I have a bad feeling about that.  The only compelling substantiation I can provide for that bad feeling is that we aren't doing it as an idiom:
martind@duezer:~/software.jessies.org/work$ sgrep '\bexit[^!]' | grep '\.rb:' | wc -l
23
martind@duezer:~/software.jessies.org/work$

We ought to be able to pay once for this wrinkle and not have to think about it elsewhere.  A better way to do that, than I've done here, would be to abstract-out the three copies of this exception handler and "rescue SystemExit" separately.  I don't think that's too hard either, but this is a step towards collapsing the duplicates.

If I'm going to abbreviate standard class names in identifiers, then, these days, I like to use the first two letters rather than just the first letter.  In particular, there's a lot of precedent for "Exception ex".

bin/invoke-java.rb: We can go back to plain exit() now, following the other idiom we have that "! methods are a bad idea".  I forget the motivating example for that, but then that's the idea - to be able to program effectively, without having to remember all the gotchas, by following a few pidgin rules.