bin/invoke-java.rb: Gain the custom Cygwin launcher support from Terminator and lose the support for letting the application decide on the launcher.  Lose the support for changing the dock name after we've been constructed.  That's just asking for us to do something dependent on the DEBUGGING_TERMINATOR name, only to have the variable's name changed under our feet.  A read-only accessor would be fine but we don't currently need it.
By using the same launcher for all Windows applications, we'll always use the same jvm.dll.  Previously, we might have been in a situation where Terminator worked and Evergreen didn't.  We were previously not checking the Java version on Cygwin when running anything but Terminator.  The comment explaining why wasn't quite right.

When it was originally checked-in, the Terminator version of this code used "launcher" rather than "launcherw" when DEBUGGING_TERMINATOR was set.  The check-in comment says that this allowed to us to see the Log.java output in the console subsystem window which would pop-up if run from a shortcut to ruby.exe (instead of our normal rubyw.exe).  This no longer works on my Cygwin installation, which is now running 1.5.21.  Now, regardless of the subsystem of the launcher, I see Java stdio output in my rxvt or Terminator or Evergreen errors window.  Now, regardless of the subsystem of the launcher, I don't see Java stdio output in cmd.exe or the Cygwin shortcut window or the window popped up by a shortcut to ruby.exe.

I used this as my test code:

/software.jessies.org/work/salma-hayek/.generated/i386_Cygwin/bin/launcherw -classpath c:/cygwin/software.jessies.org/work/salma-hayek/.generated/classes e/tools/JavaHpp badger