robots.txt: Elliott argues for hiding the Subversion access from Google's index:
it would be a bad thing to have the binary files indexed. (that's the  
reason we bother about the downloads, not because there's a  
particular problem with them caching our old binaries.)

plus i'm more than happy for people wanting to copy and paste our  
hard work into their proprietary programs to have to at least go to  
the effort of downloading our stuff. i no longer post fully-working  
code on my blog; just explain the idea and show the most interesting  
bit.