Makefile: http://software.jessies.org/Evergreen/ now says:
svn checkout http://software.jessies.org/svn/evergreen/trunk/ evergreen

But anyone building a .deb (for example) from that would (without this change) build an installer with the wrongly capitalized name in its description.  In the case of (at least) the Windows installer, it would affect the case of the desktop shortcut.

We could all check-out with a consistent case... or we could just make it work anyway.