universal.make: While accepting that it's quite usual for make-based systems to have race conditions if you change a source file during compilation, Elliott notes that we have an unusually long window during Java compilation, which bit him.  He suggested touching the sentinel file back to the date of the start of compilation, but I'd rather not get involved with timestamp granularity and, particularly, bash variables.
Add that to list of things I'd expect to be catered for by a replacement for make.  It should also know when a compilation hasn't completely succeeded, even if the make tool is killed without being able to clean up the targets (disk-full is more common reason that kill -9).  This sentinel file nonsense wants to be handled by the tool, so it's worth spending the effort to do it right, because the benefit will be shared by all targets.