universal.make: create bzip2-compressed .dmg files. I include the raw data behind this choice:
Here are the results of making the current .dmg files with the few sensible options amongst the many choices hdiutil(1) provides:

default
-rw-r--r--   1 elliotth  elliotth       2M May 10 22:50 evergreen/.generated/universal_Darwin/bin/evergreen-4.104.2228.dmg
-rw-r--r--   1 elliotth  elliotth       2M May 10 22:50 scm/.generated/universal_Darwin/bin/scm-2.8.2228.dmg
-rw-r--r--   1 elliotth  elliotth       2M May 10 22:51 terminator/.generated/universal_Darwin/bin/terminator-5.32.2228.dmg

UDCO
-rw-r--r--   1 elliotth  elliotth       2M May 10 22:53 evergreen/.generated/universal_Darwin/bin/evergreen-4.104.2228.dmg
-rw-r--r--   1 elliotth  elliotth       2M May 10 22:54 scm/.generated/universal_Darwin/bin/scm-2.8.2228.dmg
-rw-r--r--   1 elliotth  elliotth       2M May 10 22:54 terminator/.generated/universal_Darwin/bin/terminator-5.32.2228.dmg

UDZO zlib-level=1
-rw-r--r--   1 elliotth  elliotth       2M May 10 22:55 evergreen/.generated/universal_Darwin/bin/evergreen-4.104.2228.dmg
-rw-r--r--   1 elliotth  elliotth       2M May 10 22:55 scm/.generated/universal_Darwin/bin/scm-2.8.2228.dmg
-rw-r--r--   1 elliotth  elliotth       2M May 10 22:55 terminator/.generated/universal_Darwin/bin/terminator-5.32.2228.dmg

UDZO zlib-level=9
-rw-r--r--   1 elliotth  elliotth       2M May 10 22:59 evergreen/.generated/universal_Darwin/bin/evergreen-4.104.2228.dmg
-rw-r--r--   1 elliotth  elliotth       1M May 10 22:59 scm/.generated/universal_Darwin/bin/scm-2.8.2228.dmg
-rw-r--r--   1 elliotth  elliotth       2M May 10 22:58 terminator/.generated/universal_Darwin/bin/terminator-5.32.2228.dmg

UDBZ
-rw-r--r--   1 elliotth  elliotth       1M May 10 22:56 evergreen/.generated/universal_Darwin/bin/evergreen-4.104.2228.dmg
-rw-r--r--   1 elliotth  elliotth       1M May 10 22:56 scm/.generated/universal_Darwin/bin/scm-2.8.2228.dmg
-rw-r--r--   1 elliotth  elliotth       2M May 10 22:57 terminator/.generated/universal_Darwin/bin/terminator-5.32.2228.dmg

UDBZ (bzip2) has the disadvantage of only being available in Mac OS 10.4 and later, but we already require Java 5, which limits us to 10.4 or later anyway. This change will make life worse for 10.3 users, though: they won't even be able to get as far as double-clicking and getting our little error message about needing to upgrade. I don't know what 10.3 does when given a .dmg it doesn't understand, but I imagine it's not pretty.

Still, Mac users tend to upgrade swiftly, so I don't think this is a problem in practice. I'd have gone for UDZO with zlib-level=9 if it had performed as well, because that would have been compatible with older releases, but it's not as good.

Ah... hang on... these results are misleading. Mac OS' "ls -h" only shows integers, and uses round-to-zero! Let's start again from scratch with uglier but more accurate data...

default
-rw-r--r--   1 elliotth  elliotth  2433733 May 10 23:09 evergreen/.generated/universal_Darwin/bin/evergreen-4.104.2228.dmg
-rw-r--r--   1 elliotth  elliotth  2112598 May 10 23:08 scm/.generated/universal_Darwin/bin/scm-2.8.2228.dmg
-rw-r--r--   1 elliotth  elliotth  2651413 May 10 23:08 terminator/.generated/universal_Darwin/bin/terminator-5.32.2228.dmg

UDCO
-rw-r--r--   1 elliotth  elliotth  2409338 May 10 23:07 evergreen/.generated/universal_Darwin/bin/evergreen-4.104.2228.dmg
-rw-r--r--   1 elliotth  elliotth  2174589 May 10 23:07 scm/.generated/universal_Darwin/bin/scm-2.8.2228.dmg
-rw-r--r--   1 elliotth  elliotth  2836274 May 10 23:07 terminator/.generated/universal_Darwin/bin/terminator-5.32.2228.dmg

UDZO
-rw-r--r--   1 elliotth  elliotth  2324225 May 10 23:06 evergreen/.generated/universal_Darwin/bin/evergreen-4.104.2228.dmg
-rw-r--r--   1 elliotth  elliotth  2112597 May 10 23:06 scm/.generated/universal_Darwin/bin/scm-2.8.2228.dmg
-rw-r--r--   1 elliotth  elliotth  2786125 May 10 23:05 terminator/.generated/universal_Darwin/bin/terminator-5.32.2228.dmg

UDZO zlib-level=9
-rw-r--r--   1 elliotth  elliotth  2202219 May 10 22:59 evergreen/.generated/universal_Darwin/bin/evergreen-4.104.2228.dmg
-rw-r--r--   1 elliotth  elliotth  2066401 May 10 22:59 scm/.generated/universal_Darwin/bin/scm-2.8.2228.dmg
-rw-r--r--   1 elliotth  elliotth  2671822 May 10 22:58 terminator/.generated/universal_Darwin/bin/terminator-5.32.2228.dmg

UDBZ
-rw-r--r--   1 elliotth  elliotth  1958677 May 10 23:04 evergreen/.generated/universal_Darwin/bin/evergreen-4.104.2228.dmg
-rw-r--r--   1 elliotth  elliotth  2019296 May 10 23:04 scm/.generated/universal_Darwin/bin/scm-2.8.2228.dmg
-rw-r--r--   1 elliotth  elliotth  2522046 May 10 23:05 terminator/.generated/universal_Darwin/bin/terminator-5.32.2228.dmg

So there's not all that much in it. We save between 100KiB and 500KiB. That's better than nothing, but not nearly as good as I'd hoped for. Still, bzip2 is the clear winner in size terms, and I don't think we lose much by further excluding 10.3 users. The download link already says "For Mac OS 10.4".

The two highest compression choices (UDBZ and UDZO zlib-level=9) felt noticeably slower than the others, but I've ignored time in all this, because we don't build the .dmg often enough (or interactively enough) for that to be a consideration.

BTW, I'd been using du(1) this morning, and "du -h" does provide more useful (non-integer) results.