www/faq.html: add a FAQ, for consistency.
www/index.html: add a tag line, for consistency.

www/cygwin-problems.html: try to improve some upsetting grammar (which wasn't wrong, but was questionable and inconsistent). Add a note to the existing "why Cygwin?" section that was the main difference between this and the new "why Cygwin?" section I wrote in the FAQ, before remembering that we'd been here before.