lib/installer.wxs: Run Ruby using ruby-launcher, in order to unset RUBYOPT and to give Ruby valid stderr (and stdout) file descriptors.  We need to start in the Cygwin bin directory in order that ruby-launcher.exe can find cygwin1.dll.  That means we need to give a full path to the (eg) bin/terminator script, which means the path will be in the Windows style - with spaces in the path and drive letters and backslashes.  The shortcut executor seems to do the right thing with argv (coping with the spaces and quotes) and Cygwin Ruby (or cygwin1.dll) fortunately copes with the drive letters and backslashes.
I failed to find a way of reusing the existing DirectorySearch object (the help for DirectorySearchRef is impenetrable and there are no examples in the tutorial).  Note the change in the Depth attribute.  We need a depth of one here or the shortcut's working directory silently becomes empty (while still passing Orca validation).