www/downloads.html: mad reports:
  "Windows x86" site:microsoft.com gets a lot of hits, including in the
  context of being "the opposite of" x64.

so let's try to subtly hint that our .msi is 32-bit without being too verbose about it (and especially not getting in to the whole "everything but Terminator will be fine on x86 or x64, and even Terminator will work if you're using an x86 JVM on your x64 OS"). I still want to make this section less daunting, since we've done so much work to make installation "just work".