src/e/ptextarea/PAnchorSet.java: Fixed 'removeAll' so it doesn't delete PAnchors we didn't mean to delete.  Because the set's '.contains' method was used to determine whether each anchor was a deletion candidate, which is based in turn on the .equals() method of the PAnchor, which is implemented purely in terms of the anchor's index in the text buffer, the attempted deletion of any PAnchor with a text buffer index of <n> was deleting *all* PAnchors with the same text buffer index.  This was causing the weird wandering spelling mistakes bugs we were seeing when a find was cancelled.
I've implemented it using the 1.5 JDK methods; when we move to 1.6, we should use a 'NavigableSet' and use the reverse iterator available from that interface, rather than this rather ugly implementation which requires us to take a copy of the anchors to delete, and then reverse the list.

I've also added a little comment pointing out an inefficiency.  Probably not such a dangerous one, mind.