I've got another silent problem starting my shortcuts today.  If I hack the shortcuts to say ruby.exe instead of ruby-launcher rubyw, then I see a monolog box saying:
ruby.exe - Entry Point Not Found
The procedure entry point ruby_init_stack could not be located in the dynamic link library cygruby18.dll.

This is even after a reboot but, at the moment, I'm less concerned about my broken shortcuts and more concerned about the fact that the failure is silent.

native/all/ruby-launcher/ruby-launcher.cpp: Before the exec, call the suggestion from...

native/all/launcher/launcher.cpp: ... the string literal here, which can now become...

native/Headers/WindowsDllErrorModeChange.h: ... code.

I'm wary of turning it on by default in java-launcher because the posix behavior may be what saves us if and when we install both i386 and amd64 versions.  I'm checking this in to record it in the history but also to help me get the code from the editor on Linux to my Windows machine for compilation.  If it successfully reveals my ruby-launcher error and doesn't get in the way normally, then perhaps I should make the java-launcher change too.