src/terminator/terminal/TerminalControl.java: The SingleThreadExecutor we use for working around a problem with Linux 2.4 means that any bugs in how we call the JNI code can easily lock up the whole Terminator process.  We already log some output when waitFor returns but, if it doesn't return, you can be left at a loss as to why not, even if you've discovered the sudo killall hsdbd; sudo jstack nonsense that's necessary on Mac OS to extract the less informative (no thread names) version of the thread dump.  So add to the log spam to point the finger.
The instance of this bug that I tripped over happened when the reader ran out of heap space.  Flag that in the source code, where we start to go down a code path that isn't necessarily right, though handleProcessTermination may be the expedient place to fix it.  When considering whether we want another early return, I see that we're not cleaning up the writerExecutor in the case we take an early return.  That doesn't look right, so it gets a flag too.

I was actually glad that Java ran out of memory here, though it didn't stop the previously unheard of degree of fan racket from the normally nearly silent Mac Book.  I thought I'd killed the ruby script doing continuous output that I'd started in order to try, unsuccessfully, to reproduce the increasingly annoying cross-platform bug with Copy (whereby Copy sometimes seemingly does nothing).  I hadn't and, by the time we came back from the Taiwanese bakery with a delicious banana and walnut cake, it was good bye Terminator with event dispatch hangs lasting several seconds.  Killing the ruby script finally stopped the noise.

Right, so do we try to fix this in nativeWaitFor, or do we trap a more specific exception or do we abandon that work-around or make it specific to Linux or what?  The first option sounds most robust to me, so I'm off to work out how we'd do that.