native/all/pty/terminator_terminal_PtyProcess.cpp: The only place we currently call nativeWaitFor is when we've given up reading from the pipe to the child but that doesn't necessarily mean that the child has died.  My motivating case was when Terminator ran out of heap due to a script continuously outputting but another case would be that a child goes into daemon mode and closes all of its standard i/o descriptors.  WNOHANG should stop us tying up the single thread we use to run JNI code (for Linux 2.4 compatibility) providing that it compiles on all the platforms we support.  It works on Linux and compiles on Mac OS.  I'll try Cygwin when I've checked-in.
src/terminator/terminal/TerminalControl.java: handleProcessTermination has several early returns.  Shutting down the writer thread is the least important thing we're doing but we should nevertheless make an effort to do it in each code path.  (If we're worried about it throwing exceptions and stopping us from executing something else, then we'll have to trap them, but I'm not sure we really are worried.)  The pseudo-switch on the exit status was previously complete except for the unlikely case that WIFSTOPPED, which can only happen if "the child is being traced".  With WNOHANG, though, we might well return when the child hasn't exited normally and when it hasn't been signaled, so we now need to handle that case.  A more informative message might be useful but I hope no-one will ever see this case anyway.  (It might be unlikely but locking up the whole Terminator process is such an unfortunate consequence that I think the expected damage is worth trying to fix, particularly given that it's not really costing us much extra code or complexity.)