native/all/pty/terminator_terminal_PtyProcess.cpp: That didn't work reliably on Cygwin.  Perhaps it wouldn't work reliably elsewhere.  Revert to blocking until the child exits...
src/terminator/terminal/PtyProcess.java: ... after having the insight that we could use a different single-threaded executor for each child.  This still gives us the Linux 2.4 work-around but allows us to block this thread without stopping us from creating new children or waiting for other children.  The gotcha is that we have to remember to stop the executor or else we leak (daemon) threads.

This would let us remove the "contact lost" clause and the "result == 0" check of the returned value from waitpid.  But if we did ever hit that "child being traced" case and the window just disappeared, we'd kick ourselves.

Anyway, time to check this on Cygwin...