src/e/edit/TagReader.java: treat global variables like fields and unions like structs when sorting. Treat file-scope identifiers as private for the purposes of coloring. Use || instead of | in a context where it's correct but looks strange.
src/e/edit/TagType.java: add two new shapes, "diamond" and "hash", and use the latter for macro definitions. (It's a bit C-centric, but meaningful shapes are hard to come by!) Use the "triangle" that we've always used for fields for global variables too.

src/e/edit/TagsPanel.java: always set an icon on the tree cell renderer, and let the icon decide whether or not to draw anything. This moves the null test so the code's more obviously correct. The Shape draw/fill code gets even worse to cope with the new "hash". Each Shape probably needs a "canFill" boolean, and we should either draw or fill (but never both) based on that and tag.isAbstract.