src/terminator/model/TextLine.java: make it possible to pass in a "width hint" so that the result contains a larger number of shorter segments...
src/terminator/view/TerminalView.java: ...so we can manually "clip" to that granularity, and avoid the cost of trying to render an arbitrary amount of text. 200 200KiB lines, for example, goes from 4 minutes down to 1 minute on one x86 machine I tested on. The next step is probably to change TextLine.getStyledTextSegments to return an iterator rather than a collection, so we don't actually do any more work than we need to.