bin/find-jdk-root.rb: universal.make: Adapt to the new java-launcher name in comments...
bin/invoke-java.rb: ... and in the code that specifies the executable's path.  I considered changing @launcher and friends to @java_launcher.  That might still be a good idea but it obscured the necessary change so it would be better in a separate, cosmetic-only check-in.  Plain "@launcher" is probably fine in a local variable anyway - which launcher we mean is clear from context unless you're grepping and even then it's clear if you know the code.