src/e/edit/ShowHideTagsAction.java: "symbols" is less of an implementation detail than "tags", and probably more meaningful to more people. I hope "tree" might help people understand which bit it is we're talking about too, in the absence of some kind of menu-item rollover action where we shade out the rest of the UI and let just the tags panel shine through (which I think would be the best solution, but not the best use of my time).
src/e/edit/FindInFilesDialog.java: improve a tags-related error message (even though no-one but us will see it).

src/e/edit/Evergreen.java: we don't have a simple way to upgrade old XML files (the availability of which would be a convincing selling point for that nonsense), so I'll have to leave the stored configuration with the ambiguous name, but there's no reason for the identifier in the source to be "splitPaneDividerLocation" instead of "tagsPanelSplitPaneDividerLocation".