simple.make: Move the "filter" invocation onto the front of the command line, passing the universal.make command line as arguments.
lib/build/filter-build-output.rb: Run ARGV down a pipe and filter its output.

This was my first thought.  It doesn't solve the original problem - the exit status of the child command is still ignored.  I'm not sure we're filtering the command's stderr.  I think this is the right direction, so I'll check-in (as it's still basically functional) and then work on fixing those.