src/e/edit/Evergreen.java: make it clearer that "remove workspace" doesn't touch disk, which is more ambiguous now...
src/e/edit/WorkspaceProperties.java: ..."add workspace" can optionally create the directory for you if it doesn't already exist (optionally because you might just have typed it wrong). Also choose a sensible default name based on the root, for those cases where the user just types a root directory. (If they chose a root directory with the "Browse..." button, we already visibly did this for them, but if they just typed in the path themselves, we silently gave the workspace the empty title.)