src/e/edit/KillErrorsAction.java: src/e/edit/ShellCommand.java: src/e/edit/StreamMonitor.java: pass through to EErrorsWindow information about whether we're giving it text from standard out or standard error. Internal errors count as standard error. I've also improved some of the names to be more specific.
src/e/edit/EErrorsWindow.java: stop auto-scroll on error. This stops scrolling after the first "chunk" of error text, which is often just one line. That's not really what we want, but this is a proof of concept more than anything else. If it turns out to be annoying, it doesn't have to stay. (At least one person has asked for this functionality to come back recently.) Really, we'd like to be able to color the error text differently, but PTextArea's not really set up for that.