native/all/setsid: Not for the first time, this program confuses me.  It doesn't call setsid and, according to the history, it never has.
native/all/setpgid: It calls setpgid instead.  setsid is not a useless function - we use it in Terminator.  The util-linux setsid that we're hiding actually does call setsid.  We shouldn't pretend to be what we're not.

native/all/setsid/setsid.cpp: This needs replacing too, but Subversion won't let me do that until I've committed this.  Sorry.