simple.make: Move simple.make to lib/build/ too and replace it with a symlink to the new location.  I could replace it with a regular make file but then that would have to temporarily gain the make version check and a third copy of the salma-hayek locator code.  That's too much pain for a temporary crutch.
lib/build/simple.make: The disadvantage of a symlink is that we need a temporary hack in the salma-hayek locator to cope with being included through a different path.  Update the instructions too.

www/index.html: And the example error messages.

lib/build/make-new-project.sh: New projects should use the new location.

Makefile: Set a good example by using the new location ourselves.

Subversion appears unwilling to let me replace a regular file with a symlink in one check-in.  Let's see if it will let me do it in two.