src/terminator/InfoDialog.java: We're telling the log writer to suspend logging, not just to set some state variable.
src/terminator/LogWriter.java: We have this NullWriter thing, so why the blazes do we still have writer != null checks and, as Sam Bayer demonstrates, null pointer exceptions?  Let's have the invariant that writer != null, and stick to it, removing spurious checks and one double assignment to writer (firstly to null and then to NullWriter.INSTANCE, for no obvious reason).  Rename setSuspended per the InfoDialog change and rename the parameter so it doesn't look like we're testing whether we're already suspended.  Try to simplify things by collapsing as many of the references to null as possible.