src/terminator/terminal/TerminalControl.java: I don't see how we can get any output from the terminal before we create the InputStreamReader, so we can defer the creation of the LogWriter until we know the pty name which, as Sam Bayer points out, might be usefully written to the application log or used in the terminal log name.
src/terminator/LogWriter.java: Writing the pty name and the terminal log name to the application log is clearly a good idea.  And the suggester doesn't even want logging!

Elliott suggests that we might want to put the timestamp before the command line, so restrict the term "prefix" to the method body that decides to use it as a prefix (and avoid the misleading impression in the calling code that the argument isn't escaped).