src/terminator/LogWriter.java: we only need one DateFormatter for the whole of Terminator, and we never change it, so it should be "static final". We can also manage a better name; "LOG_TIMESTAMP_FORMATTER" was my first choice, but that's not as clear as "FILENAME_TIMESTAMP_FORMATTER" in the context of a class that's all about logging, where there's only one interpretation of "filename" but where "log" could refer to the filename or the file's contents. Also switch to the ISO-approved "T" separator between the date and time portions; I appear to read dates right-to-left, judging by how often I'd read the hour part of the time as the day part of the date. Finally (for this single line) add a comment explaining why we don't just use our usual ISO date formatter.
The word "stem" in filename contexts usually refers to the leafname without an extension, so let's say "leafname" where we mean leafname. We may as well have the helper put together the whole filename, too, and return it as the appropriate Java object --- the ever-misleading File.

Finally -- for this patch -- add a comment explaining why there's so much code for such a seemingly simple task.