lib/build/universal.make: One might reasonably use the nightly-build.rb script with the "install" or "installer" targets instead of native-dist.  Extend the convenience hack that says "yeah, we know you can't upload a binary distribution for salma-hayek but, if you're asking for that, you'd probably like it built".  On the one hand, this would be better in the nightly-build.rb script but, on the other, it is convenient to be able to build without remembering a special case for salma-hayek.
Given that, these days, we know when the Java definitely doesn't need building (and we've always known when the C++ doesn't need building), we could have the other projects build salma-hayek.  We've got code for building multiple, independent C++ projects from one invocation of make, so how hard could extending that to Java be?  It looks significantly simpler, mind, to get simple.make to recurse into salma-hayek first.  But then we'd still want this hack, or something functionally equivalent.