www/index.html: We don't compile with WiX v3.  It's hard to know whether it would be a moving target while it's still in beta.  The documentation doesn't seem to be there.
One problem that I encountered while trying WiX 3 was that my .NET framework wasn't up-to-date enough.  That would have been no biggie except for the way Cygwin causes exec(2) to fail with silent "success" when an executable which requires an absent DLL is run.  Perhaps we should run candle.exe via ruby-launcher, which already makes the necessary SetErrorMode invocation.  It would be a perversion of the name but it might just do the trick.  Sadly, of course, I can no longer easily test this.  Perhaps I should try installing WiX 3 at work (but not over my WiX 2 installation) and see if it fails the same way there.

I would guess that we don't need MergeMod.dll since we stopped building salma-hayek as a WiX module.  Maybe I should find out for sure but this machine is slow and Windows remains distressingly opaque, so it's probably easier to wait until someone needs to install WiX on a Windows box where it isn't already installed.