native/all/pty/PtyGenerator.h: native/all/pty/terminator_terminal_PtyProcess.cpp: offer an interface more like the underlying execvp(3) call, with a separate executable name and initial argv, rather than assuming that argv[0] is the executable name.
src/terminator/terminal/PtyProcess.java: offer the same interface on the Java side.

src/terminator/view/JTerminalPane.java: move the code that attempts to ensure a login shell...

src/terminator/terminal/TerminalControl.java: ...down a level. This lets us use the 1970s trick of setting argv[0] to "-" when starting a shell to inform it that it's a login shell. Why? Because tcsh(1) is so broken that "-l" (which we were already using instead of Bash's "--login" for portability's sake) can only appear on its own.

src/terminator/LogWriter.java: while touching all this code, I think it makes sense to pass the List<String> around further (because that's the newer idiom), especially now the String[] we have conveniently available will be mangled into unsuitability for the LogWriter's purposes in most cases.