bin/invoke-java.rb: Gain the code from Evergreen to stop Cygwin's /etc/profile from scuppering the effect of Terminator's --working-directory parameter.  We should set this variable here so that it takes effect when the user starts Terminator from the command line as well as when it's started from the desktop shortcut.
There's another piece to this puzzle - changing to the user's home directory when starting applications from the desktop shortcut.  I'm not sure what's the best way to do that.  While I prevaricate, double-clicking on Terminator in Windows will leave you in /usr/bin.  At least that's obvious from the prompt.  I've gotta dash, but checking in is a convenient way of getting the code from my MacBook to my Windows box - and will force me to finish this off rather than letting it slide.