lib/build/installer.wxs: Braden Anderson has a repeatable problem with Terminator trying to install itself into C:\CFusionMX7\bin.  A dump of his registry revealed that he had stub entries for the Cygwin registry keys under HKCU.  A look in my registry revealed that so did I - a vestige of a per-user installation.  A look at msiexec /l* install.log /i terminator.msi showed that the CYGWIN_HKCU_BIN property was being set even though CYGWIN_HKCU wasn't.  It seems like it was just searching for the first "bin" directory that came to hand.  Looking at the same log generated on my system showed that mine was setting CYGWIN_HKCU_BIN too, fortunately (or otherwise) to C:\cygwin\bin.
Fix that by not using "[property]" in Path attributes.  Instead we'll have to duplicate the parent searches.  I tried again to use DirectorySearchRef but you know you're on a hiding to nothing when the most useful of the top ten Google results you turn up is one of your own check-in comments berating the lack of examples and impenetrability of this feature.

We're now using the CYGWIN_HKLM_BIN property in only two places.  One is in a Condition element, where it was added under r1705 with this check-in comment [

The difficult part was finding that this perturbed evaluation order of the CYGWIN and RUBYW properties such that, even though the RUBYW property uses the CYGWIN property, it was now evaluated first - and thus didn't work.  What a half-baked little language.

].  As well as exhibiting even less understanding of the Windows installer structure than I now have, I can't see how that can still be true after the change I'm making here.

The other usage is in setting the working directory for the shortcut.  That's something we could avoid the need for if I write the cygwin-launcher.exe I was mulling the other day (under the auspices of the "Terminator Here" RFE).  Ah, but then I'd want to pass the Cygwin bin directory to the cygwin-launcher, so I won't be able to get rid of this duplication even then.  Oh well.

This installs OK and passes Orca validation - and doesn't assign properties that it shouldn't.  Whether it will work for Braden is another matter.  I'm not sure it will because his installer log says it bound C:\CFusionMX7\bin to both CYGWIN_HKCU_BIN *and* CYGWIN_HKLM_BIN, though it bound CYGWIN_HKLM and RUBYW_HKLM correctly.  Previously it was alleging a successful installation but starting the shortcut in the wrong directory.