lib/build/universal.make: Chris's Sparc Solaris scp produces this output when we should upload a new pkg file:
scp: warning: stat: SJOevergreen_5.1.2527_sun4u.pkg (src): no such file (server msg: 'No such file')

The unfortunate use of "warning: " in that *error* message matches a regular expression in filter-build-output.rb, causing it to reason "looks like a Java warning - better output that and the preceding commands".  We're already paying for an ssh connection to do a mkdir -p, so it's not much hassle to avoid the error.  Or warning.