src/terminator/LogWriter.java: We only use ptyName in a Log.warn.  Supplying it has caused a problem to our one caller.  We expose all the information that our caller would need in order to do the Log.warn itself.  Let's do that to stop the caller having to jump through exceptional hoops.
src/terminator/terminal/TerminalControl.java: We no longer need to explicitly use the pty name at all, as it's included in the ptyProcess.toString() output.  We will now indicate in the application log that we've been unable to open the terminal log because the logs directory is unwritable.  That sounds like an improvement to me.  Now we definitely need a code-comment.