native/all/pty/PtyGenerator.h: http://www.win.tue.nl/~aeb/linux/lk/lk-10.html#ss10.3 confirms my suspicion, which had already been reinforced by comments in an earlier version of our source ("44BSD way to acquire controlling terminal"), that the TIOCSCTTY ioctl is the BSD approach for acquiring the controlling terminal.  Linux, it confirms, tries to support both models and should work fine with our current code.  Cygwin's source only appears to support the System V approach.  Another early comment suggests it would be a bad idea to call TIOCSCTTY on "SunOS" but we seem to be getting away with it on Solaris.
I've now had a look at the libvte4 source and seen that, like our current code, it doesn't use O_NOCTTY when opening the slave tty in the child process.  It uses it on the master open, as we do.  At the time of the change that first introduced O_NOCTTY to our source, we didn't pass it to the master open.  That was fixed when we switched to posix_openpt.  Perhaps O_NOCTTY was simply added to the slave open when the intention was to add it to the master.  I doubt we'll be able to find out this many years later but I'm now more confident that my changes of yesterday aren't going to be a regression.