lib/build/universal.make: C_AND_CXX_FLAGS is only evaluated during the execution of the compilation rule.  At that point, it's an error to refer to $(SOURCE_DIRECTORY).  SOURCE_DIRECTORY should be treated just like a per-directory local variable.  Think of it as the parameter to that scope - parameters are clearly local.  Add SOURCE_DIRECTORY to the list of LOCAL_VARIABLES, so we get the benefit of a early failure with a half-decent error message.
That simple change caused a knock-on problem with BASE_NAME.  We had a special hack allowing that local variable to be used outside the local scope for the benefit of the scoping violation error message.  But BASE_NAME is, it seems, defined in terms of SOURCE_DIRECTORY, which is now a local variable.  For those error messages, I think we only need to know the name of the immediately preceding scope.  We know where we want the name of the immediately preceding scope and so there we can ask for it with a special name.  That removes the possibility that someone will use BASE_NAME somewhere else and get misleading results with no protection.  I clarified the error message while I was there, using more consistent quoting.

Now, elsewhere, we add components to C_AND_CXX_FLAGS which depend on $<.  So I can follow that pattern and make the -mno-cygwin compilation flag dependent on that instead of SOURCE_DIRECTORY.  That's fine, as far as it goes, but I can't use that for LDFLAGS because per-directory.make treats that variable quite differently.  It's evaluated at the start of per-directory.make and then added to according to whether we find Cocoa source and so on.  That seems to have lasted OK under maintenance too but, to fit in with it for Mingw, I'm going to have to move the LDFLAGS setting...

lib/build/per-directory.make: ... here.  This smacks of distributed duplication and I'm not very happy about it but I don't want to bite off some large-scale reorganization while I'm in the middle of the cygwin-launcher feature.

Before this change, cygwin-launcher was linked with -mno-cygwin but compiled without that flag, despite there being (non-functioning) code to enable that flag for the compilation.