native/Mingw/cygwin-launcher/cygwin-launcher.cpp: A (blech) copy-and-paste of the parts of ruby-launcher.cpp that we want here, namely the WindowsDllErrorModeChange and the message box error handling for the case where we can't start the Cygwin program.  The only changes are in boilerplate text, a removal of #include directives that we don't need here and the addition of a TODO for the one extra, motivating, thing that we want to do here.
When the previous check-in's change started running the compiler with the right flag, it became clear that the Microsoft runtime doesn't provide strerror_r.  For the purposes of cygwin-launcher...

native/Headers/strerror_r.h: ... we can provide a trivial implementation in terms of strerror...

native/Headers/unix_exception.h: ... that does enough to satisfy the caller here.