native/Headers/reportFatalErrorViaGui.h: cygwin-launcher wasn't going down the MessageBox route because __CYGWIN__ isn't defined for Mingw builds.
native/Headers/WindowsDllErrorModeChange.h: This will also affecting cygwin-launcher (though I haven't tested it).

native/Mingw/cygwin-launcher/cygwin-launcher.cpp: A lengthy investigation of the history of this now obviously incomplete stanza for including windows.h leads me to conclude that if and when we really do need to include it first, we'll see a compilation error.  I'll treat it as a normal #include here.  I need no OS guard because this directory is for Windows-specific source.  I will eventually need windows.h for the SetDllDirectory call.